scide: Split main() out in to separate file
