+++ editors/sc-ide/CMakeLists.txt
@@ -126,6 +126,7 @@ file (GLOB_RECURSE all_hdr *hpp)
 
 set ( ide_src
     core/main.cpp
+    core/main_function.cpp
     core/app_palette.cpp
     core/sig_mux.cpp
     core/doc_manager.cpp
