+++ editors/sc-ide/core/main.cpp
@@ -47,90 +47,6 @@
 
 using namespace ScIDE;
 
-int main(int argc, char* argv[]) {
-    QApplication app(argc, argv);
-
-    QStringList arguments(QApplication::arguments());
-    arguments.pop_front(); // application path
-
-    // Pass files to existing instance and quit
-
-    SingleInstanceGuard guard;
-    if (guard.tryConnect(arguments))
-        return 0;
-
-    // Set up translations
-    QTranslator qtTranslator;
-    qtTranslator.load("qt_" + QLocale::system().name(), QLibraryInfo::location(QLibraryInfo::TranslationsPath));
-    app.installTranslator(&qtTranslator);
-
-    QString ideTranslationPath = standardDirectory(ScResourceDir) + "/translations";
-
-    bool translationLoaded;
-
-    // Load fallback translator that only handles plural forms in English
-    QTranslator fallbackTranslator;
-    translationLoaded = fallbackTranslator.load("scide", ideTranslationPath);
-    app.installTranslator(&fallbackTranslator);
-    if (!translationLoaded)
-        qWarning("scide warning: Failed to load fallback translation file.");
-
-    // Load translator for locale
-    QString ideTranslationFile = "scide_" + QLocale::system().name();
-    QTranslator scideTranslator;
-    scideTranslator.load(ideTranslationFile, ideTranslationPath);
-    app.installTranslator(&scideTranslator);
-
-    // Force Fusion style to appear consistently on all platforms.
-    app.setStyle(QStyleFactory::create("Fusion"));
-
-    // Palette must be set before style, for consistent application.
-    Main* main = Main::instance();
-    main->setAppPaletteFromSettings();
-
-    // Install style proxy.
-    app.setStyle(new ScIDE::Style(app.style()));
-
-    // Go...
-    MainWindow* win = new MainWindow(main);
-
-    // NOTE: load session after GUI is created, so that GUI can respond
-    Settings::Manager* settings = main->settings();
-    SessionManager* sessions = main->sessionManager();
-
-    // NOTE: window has to be shown before restoring its geometry,
-    // or else restoring maximized state will fail, if it has ever before
-    // been saved un-maximized.
-    win->show();
-
-    QString startSessionName = settings->value("IDE/startWithSession").toString();
-    if (startSessionName == "last") {
-        QString lastSession = sessions->lastSession();
-        if (!lastSession.isEmpty()) {
-            sessions->openSession(lastSession);
-        }
-    } else if (!startSessionName.isEmpty()) {
-        sessions->openSession(startSessionName);
-    }
-
-    if (!sessions->currentSession()) {
-        win->restoreWindowState();
-        sessions->newSession();
-    }
-
-    foreach (QString argument, arguments) { main->documentManager()->open(argument); }
-
-    win->restoreDocuments();
-
-    bool startInterpreter = settings->value("IDE/interpreter/autoStart").toBool();
-    if (startInterpreter)
-        main->scProcess()->startLanguage();
-
-#ifdef SC_USE_QTWEBENGINE
-    HelpBrowserWebSocketServices hbServices(win->helpBrowserDocklet()->browser());
-#endif
-    app.exec();
-}
 
 bool SingleInstanceGuard::tryConnect(QStringList const& arguments) {
     const int maxNumberOfInstances = 128;
