+++ editors/sc-ide/core/main_function.cpp
@@ -1,0 +1,133 @@
+/*
+    SuperCollider Qt IDE
+    Copyright (c) 2012 Jakob Leben & Tim Blechmann
+    http://www.audiosynth.com
+
+    This program is free software; you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation; either version 2 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program; if not, write to the Free Software
+    Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301  USA
+*/
+
+#include "main.hpp"
+#include "settings/manager.hpp"
+#include "session_manager.hpp"
+#include "util/standard_dirs.hpp"
+#include "../widgets/main_window.hpp"
+#include "../widgets/lookup_dialog.hpp"
+#include "../widgets/code_editor/highlighter.hpp"
+#include "../widgets/style/style.hpp"
+#include "../../../QtCollider/hacks/hacks_mac.hpp"
+#include "../primitives/localsocket_utils.hpp"
+
+#include <yaml-cpp/node/node.h>
+#include <yaml-cpp/parser.h>
+
+#include <QAction>
+#include <QApplication>
+#include <QBuffer>
+#include <QDataStream>
+#include <QDir>
+#include <QFileOpenEvent>
+#include <QLibraryInfo>
+#include <QTranslator>
+#include <QDebug>
+#include <QStyleFactory>
+
+#include "util/HelpBrowserWebSocketServices.hpp"
+
+using namespace ScIDE;
+
+int main(int argc, char* argv[]) {
+    QApplication app(argc, argv);
+
+    QStringList arguments(QApplication::arguments());
+    arguments.pop_front(); // application path
+
+    // Pass files to existing instance and quit
+
+    SingleInstanceGuard guard;
+    if (guard.tryConnect(arguments))
+        return 0;
+
+    // Set up translations
+    QTranslator qtTranslator;
+    qtTranslator.load("qt_" + QLocale::system().name(), QLibraryInfo::location(QLibraryInfo::TranslationsPath));
+    app.installTranslator(&qtTranslator);
+
+    QString ideTranslationPath = standardDirectory(ScResourceDir) + "/translations";
+
+    bool translationLoaded;
+
+    // Load fallback translator that only handles plural forms in English
+    QTranslator fallbackTranslator;
+    translationLoaded = fallbackTranslator.load("scide", ideTranslationPath);
+    app.installTranslator(&fallbackTranslator);
+    if (!translationLoaded)
+        qWarning("scide warning: Failed to load fallback translation file.");
+
+    // Load translator for locale
+    QString ideTranslationFile = "scide_" + QLocale::system().name();
+    QTranslator scideTranslator;
+    scideTranslator.load(ideTranslationFile, ideTranslationPath);
+    app.installTranslator(&scideTranslator);
+
+    // Force Fusion style to appear consistently on all platforms.
+    app.setStyle(QStyleFactory::create("Fusion"));
+
+    // Palette must be set before style, for consistent application.
+    Main* main = Main::instance();
+    main->setAppPaletteFromSettings();
+
+    // Install style proxy.
+    app.setStyle(new ScIDE::Style(app.style()));
+
+    // Go...
+    MainWindow* win = new MainWindow(main);
+
+    // NOTE: load session after GUI is created, so that GUI can respond
+    Settings::Manager* settings = main->settings();
+    SessionManager* sessions = main->sessionManager();
+
+    // NOTE: window has to be shown before restoring its geometry,
+    // or else restoring maximized state will fail, if it has ever before
+    // been saved un-maximized.
+    win->show();
+
+    QString startSessionName = settings->value("IDE/startWithSession").toString();
+    if (startSessionName == "last") {
+        QString lastSession = sessions->lastSession();
+        if (!lastSession.isEmpty()) {
+            sessions->openSession(lastSession);
+        }
+    } else if (!startSessionName.isEmpty()) {
+        sessions->openSession(startSessionName);
+    }
+
+    if (!sessions->currentSession()) {
+        win->restoreWindowState();
+        sessions->newSession();
+    }
+
+    foreach (QString argument, arguments) { main->documentManager()->open(argument); }
+
+    win->restoreDocuments();
+
+    bool startInterpreter = settings->value("IDE/interpreter/autoStart").toBool();
+    if (startInterpreter)
+        main->scProcess()->startLanguage();
+
+#ifdef SC_USE_QTWEBENGINE
+    HelpBrowserWebSocketServices hbServices(win->helpBrowserDocklet()->browser());
+#endif
+    app.exec();
+}
